requirements.txt
Flask==2.0.1
# Pillow==8.3.1
Pillow
loguru
dataclasses_json
requests
bootstrap-flask
psycopg2
flask-sqlalchemy
flask-login
flask-wtf
flask-admin
